requirements.txt
appdirs==1.4.3
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
distlib==0.3.0
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.12.1
filelock==3.0.12
idna==2.10
pipenv==2018.11.26
postgres==3.0.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
psycopg2-pool==1.1
pytz==2020.1
requests==2.24.0
six==1.14.0
sqlparse==0.4.1
urllib3==1.25.11
views==0.3
virtualenv==20.0.16
virtualenv-clone==0.5.4
faker
